added new player and started working on business logic behind processor

+++ src/internal/utils/configuration.go
@@ -1,4 +1,4 @@
 package utils
 
 const PAGE_SIZE int16 = 512
-const INITIAL_ELO int = 1300
+const INITIAL_ELO int16 = 1300
